Implementation of authentication with credential, OTP and credential with OTP

+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/entity/CredentialPolicyDetail.java
@@ -40,8 +40,8 @@ public class CredentialPolicyDetail {
     private Integer credentialLengthMin;
     private Integer credentialLengthMax;
     private String credentialAllowedChars;
-    private Integer limitSoft;
-    private Integer limitHard;
+    private Long limitSoft;
+    private Long limitHard;
     private int checkHistoryCount;
     private boolean rotationEnabled;
     private Integer rotationDays;
