+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/request/UpdateOtpPolicyRequest.java
@@ -33,7 +33,7 @@ public class UpdateOtpPolicyRequest {
     private String description;
     @NotNull
     private Integer length;
-    private Integer attemptLimit;
+    private Long attemptLimit;
     private Long expirationTime;
     private String genAlgorithm;
     private OtpPolicyStatus otpPolicyStatus;
