+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/response/CredentialAuthenticationResponse.java
@@ -41,6 +41,6 @@ public class CredentialAuthenticationResponse {
     @NotNull
     private AuthenticationResult authenticationResult;
     @NotNull
-    private Integer remainingAttempts;
+    private Long remainingAttempts;
 
 }
