+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/response/OtpAuthenticationResponse.java
@@ -44,6 +44,6 @@ public class OtpAuthenticationResponse {
     @NotNull
     private AuthenticationResult authenticationResult;
     @NotNull
-    private Integer remainingAttempts;
+    private Long remainingAttempts;
 
 }
